+++ index.js
@@ -1,21 +1,19 @@
 'use strict';
 
-const appConfig = require('./lib/config');
 const utils = require('./lib/utils');
+const defaults = require('./lib/config');
 const authCodeModule = require('./lib/client/auth-code');
 const passwordModule = require('./lib/client/password');
 const clienteCredentialsModule = require('./lib/client/client');
 const accessTokenModule = require('./lib/client/access-token');
-const coreModule = require('./lib/core');
 
-module.exports = function (config) {
-  config = utils.configure(config, appConfig);
+module.exports = function (options) {
+  const config = utils.configure(options, defaults);
 
   return {
     authCode: authCodeModule(config),
     password: passwordModule(config),
     client: clienteCredentialsModule(config),
     accessToken: accessTokenModule(config),
-    api: coreModule(config).api,
   };
 };