+++ test/auth-code.js
@@ -6,7 +6,8 @@ const nock = require('nock');
 const expect = require('chai').expect;
 const oauth2Module = require('./../index.js');
 
-const oauth2 = oauth2Module(require('./fixtures/oauth-options'));
+const oauth2 = oauth2Module
+  .create(require('./fixtures/oauth-options'));
 
 const tokenParams = {
   code: 'code',
@@ -34,20 +35,24 @@ describe('oauth2.authCode', function () {
 
   describe('#authorizeURL', function () {
     it('returns the authorization URI', function () {
-      result = oauth2.authCode.authorizeURL(authorizeConfig);
+      result = oauth2.authorizationCode.authorizeURL(authorizeConfig);
 
       const expected = `https://example.org/oauth/authorize?redirect_uri=${encodeURIComponent('http://localhost:3000/callback')}&scope=user&state=02afe928b&response_type=code&client_id=client-id`;
       expect(result).to.be.equal(expected);
     });
 
     it('should allow absolute URI for authorizationPath', function () {
-      const oauth2Temp = oauth2Module({
-        clientID: 'client-id',
-        clientSecret: 'client-secret',
-        site: 'https://example.org',
-        authorizationPath: 'https://othersite.com/oauth/authorize',
+      const oauth2Temp = oauth2Module.create({
+        client: {
+          id: 'client-id',
+          secret: 'client-secret',
+        },
+        auth: {
+          tokenHost: 'https://example.org',
+          authorizeHost: 'https://othersite.com',
+        },
       });
-      result = oauth2Temp.authCode.authorizeURL(authorizeConfig);
+      result = oauth2Temp.authorizationCode.authorizeURL(authorizeConfig);
 
       const expected = `https://othersite.com/oauth/authorize?redirect_uri=${encodeURIComponent('http://localhost:3000/callback')}&scope=user&state=02afe928b&response_type=code&client_id=client-id`;
       expect(result).to.be.equal(expected);
@@ -63,13 +68,13 @@ describe('oauth2.authCode', function () {
     });
 
     beforeEach(function (done) {
-      oauth2.authCode.getToken(tokenParams, function (e, r) {
+      oauth2.authorizationCode.getToken(tokenParams, function (e, r) {
         error = e; result = r; done();
       });
     });
 
     beforeEach(function () {
-      return oauth2.authCode
+      return oauth2.authorizationCode
         .getToken(tokenParams)
         .then(function (r) { resultPromise = r; })
         .catch(function (e) { errorPromise = e; });