+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Changed public api to, to make it consistent. Changed shortcut names to full names.
 * Changed public api to allow different sites for /authorize and /tokens
 * Added strict schema validation to module options.
+* Does not override expires_at property if passed into accessToken.create.
 
 ## v0.8.0 (1 August 2016)
 * Upgraded code to strict mode.